Example to connect two networks via routers
sets up two networks with a router each. Have a look at proj1.tf and
proj2.tf. In each network a VM is started.

A transfer network is created between the routers and ports are
assigned. As a last step, a route for the opposite network is created in
each router.